Print bone microCT data

This is a wrapper around DT::datatable() to print out bone microCT data at one site by sex for use in an R Markdown document. Each sex is displayed as an interactive, sortable, searchable table.

Usage

print_data(data, ...)

Arguments

data

A data frame containing bone microCT data, formatted as is the output of read_trabecular() or read_cortical().

...

Additional arguments passed on to DT::datatable().

Value

An htmltools::tagList() containing one DT::datatable() widget per sex.

Examples

mouse_table <- read_mouse_table(klct_ex("mouse-table.csv"))
spine_trab <- read_trabecular(klct_ex("spine_trabecular_data.csv"),
                              mouse_table, site = "Spine")
print_data(spine_trab)
#> <h4>Sex: F</h4>
#> <div class="datatables html-widget html-fill-item" id="htmlwidget-ac96cb3ee4656e2e9ec3" style="width:100%;height:auto;"></div>
#> <script type="application/json" data-for="htmlwidget-ac96cb3ee4656e2e9ec3">{"x":{"filter":"none","vertical":false,"data":[[1579,1580,1587,1632,1734,1735,1850,1854,1970,1971,2056,2057,2059,2251,2252,2255,1633,1634,1692,1693,1694,1695,1736,1846,1851,1855,1856,1968,1972,1973,2061,2064],[0.4514,0.5409,0.7058,0.3002,0.2713,0.1837,0.2674,0.2651,0.351,0.3113,0.2984,0.2544,0.2509,0.2874,0.3257,0.3054,0.4162,0.4738,0.1933,0.1067,0.2609,0.2408,0.2073,0.2419,0.3348,0.3879,0.2304,0.4693,0.4397,0.7865,0.2922,0.2575],[2.415,2.504,2.887,2.546,2.543,2.314,3.012,2.471,3.076,2.707,2.444,2.124,2.542,2.442,2.81,2.772,2.645,2.629,2.678,2.138,2.006,2.512,2.486,2.655,2.684,2.889,2.169,3.079,2.889,2.919,2.737,2.738],[18.69,21.6,24.45,11.79,10.67,7.94,8.875999999999999,10.73,11.41,11.5,12.21,11.98,9.871,11.77,11.59,11.02,15.74,18.02,7.219,4.989,13.01,9.587,8.339,9.109999999999999,12.47,13.42,10.62,15.24,15.22,26.95,10.68,9.404],[32.08,28.93,36.11,21.93,18.68,12.06,17.19,18.05,21.91,20.83,17.55,15.37,14.6,17.56,18.71,17.81,25.57,24.71,14.24,8.300000000000001,18.34,15.91,14.05,16.45,20.39,28.12,16.81,27.54,27.94,37.9,17.71,15.39],[13.28,11.55,12.51,8.615,7.345,5.212,5.708,7.305,7.123,7.696,7.179,7.238,5.742,7.191,6.661,6.426,9.669,9.401999999999999,5.315,3.882,9.140000000000001,6.335,5.652,6.193,7.598,9.731999999999999,7.748,8.945,9.670999999999999,12.98,6.47,5.62],[0.03202,0.04502,0.05262,0.03285,0.0425,0.04618,0.04585,0.03826,0.04478,0.03943,0.04606,0.04832,0.05,0.04689,0.0547,0.05125,0.04375,0.0517,0.03804,0.03701,0.0378,0.04141,0.04307,0.04327,0.0454,0.03724,0.03825,0.04239,0.03953,0.05092,0.04249,0.04884],[0.2522,0.2923,0.244,0.2924,0.4467,0.4795,0.4269,0.4156,0.412,0.4182,0.397,0.34,0.461,0.4206,0.4209,0.4363,0.3205,0.3114,0.4628,0.4691,0.2954,0.4384,0.474,0.4464,0.35,0.3651,0.3843,0.3424,0.3191,0.2696,0.4665,0.4372],[3.518,2.965,3.371,3.075,2.044,1.902,2.115,2.203,2.189,2.185,2.257,2.575,1.957,2.139,2.103,2.051,2.746,2.754,1.997,1.976,3.001,2.084,1.934,2.042,2.529,2.485,2.367,2.599,2.789,3.12,1.965,2.058],["F","F","F","F","F","F","F","F","F","F","F","F","F","F","F","F","F","F","F","F","F","F","F","F","F","F","F","F","F","F","F","F"],["AP","AP","AP","AP","AP","AP","AP","AP","AP","AP","AP","AP","AP","AP","AP","AP","Vehicle","Vehicle","Vehicle","Vehicle","Vehicle","Vehicle","Vehicle","Vehicle","Vehicle","Vehicle","Vehicle","Vehicle","Vehicle","Vehicle","Vehicle","Vehicle"]],"container":"<table class=\"display\">\n  <thead>\n    <tr>\n      <th>AS<\/th>\n      <th>Tb.BV<\/th>\n      <th>Tb.TV<\/th>\n      <th>Tb.BV/TV<\/th>\n      <th>Tb.BS<\/th>\n      <th>Tb.BS/TV<\/th>\n      <th>Tb.Th<\/th>\n      <th>Tb.Sp<\/th>\n      <th>Tb.N<\/th>\n      <th>Sex<\/th>\n      <th>Treatment<\/th>\n    <\/tr>\n  <\/thead>\n<\/table>","options":{"pageLength":25,"scrollX":true,"columnDefs":[{"className":"dt-right","targets":[0,1,2,3,4,5,6,7,8]},{"name":"AS","targets":0},{"name":"Tb.BV","targets":1},{"name":"Tb.TV","targets":2},{"name":"Tb.BV/TV","targets":3},{"name":"Tb.BS","targets":4},{"name":"Tb.BS/TV","targets":5},{"name":"Tb.Th","targets":6},{"name":"Tb.Sp","targets":7},{"name":"Tb.N","targets":8},{"name":"Sex","targets":9},{"name":"Treatment","targets":10}],"order":[],"autoWidth":false,"orderClasses":false}},"evals":[],"jsHooks":[]}</script>
#> <h4>Sex: M</h4>
#> <div class="datatables html-widget html-fill-item" id="htmlwidget-e5c8c404fe174e4c81bd" style="width:100%;height:auto;"></div>
#> <script type="application/json" data-for="htmlwidget-e5c8c404fe174e4c81bd">{"x":{"filter":"none","vertical":false,"data":[[1581,1582,1583,1584,1585,1586,1630,1690,1843,1844,1847,1848,1852,2054,2055,2058,2060,2250,2253,2254,1631,1691,1696,1733,1845,1849,1853,1967,1969,2062,2063,2065,2067,2247,2248,2249],[0.2144,0.3558,0.4205,0.2149,0.4547,0.2185,0.4102,0.2971,0.3192,0.4945,0.3289,0.2488,0.2374,0.2308,0.3356,0.2143,0.2843,0.275,0.2664,0.157,0.4033,0.3461,0.2608,0.3892,0.4652,0.2055,0.303,0.3149,0.3564,0.3108,0.2814,0.2989,0.3208,0.4165,0.267,0.304],[2.708,2.786,2.514,2.677,3.17,2.396,3.182,2.429,3.046,2.609,2.819,2.452,2.189,2.518,2.761,2.874,2.389,2.585,2.881,2.499,3.235,2.654,2.491,2.644,2.806,2.169,2.702,2.848,2.909,2.616,2.436,3.241,3.196,2.643,2.691,2.572],[7.917,12.77,16.73,8.025,14.35,9.119999999999999,12.89,12.23,10.48,18.95,11.67,10.15,10.85,9.167999999999999,12.16,7.457,11.9,10.64,9.244,6.281,12.47,13.04,10.47,14.72,16.58,9.475,11.21,11.06,12.25,11.88,11.55,9.223000000000001,10.04,15.76,9.922000000000001,11.82],[16.36,24.65,25.93,16.03,29.58,16.55,29.03,21.32,23.73,27.84,23.62,21.11,18.12,16.72,22.32,15.49,17.84,20.17,19.72,12.79,26.14,22.77,19.6,26.05,29.11,15.96,23.18,26.1,27.06,22.12,18.1,19.47,18.8,25.9,20.69,23.8],[6.042,8.845000000000001,10.31,5.988,9.332000000000001,6.907,9.124000000000001,8.776999999999999,7.793,10.67,8.379,8.609999999999999,8.279,6.641,8.085000000000001,5.391,7.466,7.803,6.845,5.117,8.082000000000001,8.576000000000001,7.869,9.853,10.37,7.357,8.577999999999999,9.164999999999999,9.302,8.457000000000001,7.431,6.009,5.881,9.798,7.688,9.255000000000001],[0.03725,0.03625,0.04401,0.03893,0.03966,0.03732,0.03665,0.03704,0.0375,0.04743,0.03672,0.03167,0.03463,0.04058,0.03975,0.04114,0.04487,0.03427,0.03824,0.03847,0.04125,0.04145,0.03476,0.04246,0.04339,0.03491,0.03511,0.03006,0.03186,0.0393,0.04381,0.0463,0.05365,0.04265,0.03479,0.03455],[0.3289,0.2727,0.2496,0.3341,0.2472,0.2988,0.2639,0.2553,0.2977,0.2323,0.2816,0.2563,0.2654,0.3699,0.2833,0.3247,0.3092,0.2829,0.2881,0.4491,0.2951,0.2791,0.2812,0.2493,0.2306,0.2875,0.2962,0.2664,0.2558,0.2847,0.2991,0.3546,0.4122,0.2477,0.3042,0.2449],[2.731,3.236,3.405,2.681,3.486,2.975,3.328,3.421,2.983,3.575,3.142,3.472,3.333,2.436,3.096,2.733,2.824,3.153,3.064,2.051,2.973,3.119,3.165,3.428,3.65,3.102,3.018,3.373,3.476,3.086,2.916,2.494,2.147,3.444,2.95,3.578],["M","M","M","M","M","M","M","M","M","M","M","M","M","M","M","M","M","M","M","M","M","M","M","M","M","M","M","M","M","M","M","M","M","M","M","M"],["AP","AP","AP","AP","AP","AP","AP","AP","AP","AP","AP","AP","AP","AP","AP","AP","AP","AP","AP","AP","Vehicle","Vehicle","Vehicle","Vehicle","Vehicle","Vehicle","Vehicle","Vehicle","Vehicle","Vehicle","Vehicle","Vehicle","Vehicle","Vehicle","Vehicle","Vehicle"]],"container":"<table class=\"display\">\n  <thead>\n    <tr>\n      <th>AS<\/th>\n      <th>Tb.BV<\/th>\n      <th>Tb.TV<\/th>\n      <th>Tb.BV/TV<\/th>\n      <th>Tb.BS<\/th>\n      <th>Tb.BS/TV<\/th>\n      <th>Tb.Th<\/th>\n      <th>Tb.Sp<\/th>\n      <th>Tb.N<\/th>\n      <th>Sex<\/th>\n      <th>Treatment<\/th>\n    <\/tr>\n  <\/thead>\n<\/table>","options":{"pageLength":25,"scrollX":true,"columnDefs":[{"className":"dt-right","targets":[0,1,2,3,4,5,6,7,8]},{"name":"AS","targets":0},{"name":"Tb.BV","targets":1},{"name":"Tb.TV","targets":2},{"name":"Tb.BV/TV","targets":3},{"name":"Tb.BS","targets":4},{"name":"Tb.BS/TV","targets":5},{"name":"Tb.Th","targets":6},{"name":"Tb.Sp","targets":7},{"name":"Tb.N","targets":8},{"name":"Sex","targets":9},{"name":"Treatment","targets":10}],"order":[],"autoWidth":false,"orderClasses":false}},"evals":[],"jsHooks":[]}</script>