Print bone microCT data

This is a wrapper around knitr::kable() to print out bone microCT data at one site by sex for use in an R Markdown document. A code chunk containing this function should probably have the results = "asis" option set to allow for pretty formatting.

Usage

print_data(data, ...)

Arguments

data

A data frame containing bone microCT data, formatted as is the output of read_trabecular() or read_cortical().

...

Additional arguments passed on to knitr::kable().

Value

Text output which by default is a Pandoc markdown pipe table for each sex contained in the data supplied.

Examples

mouse_table <- read_mouse_table(klct_ex("mouse-table.csv"))
spine_trab <- read_trabecular(klct_ex("spine_trabecular_data.csv"),
                              mouse_table, site = "Spine")
print_data(spine_trab)
#> 
#> 
#> |   AS|  Tb.BV| Tb.TV| Tb.BV/TV| Tb.BS| Tb.BS/TV|   Tb.Th|  Tb.Sp|  Tb.N|Sex |Treatment |
#> |----:|------:|-----:|--------:|-----:|--------:|-------:|------:|-----:|:---|:---------|
#> | 1579| 0.4514| 2.415|   18.690| 32.08|   13.280| 0.03202| 0.2522| 3.518|F   |AP        |
#> | 1580| 0.5409| 2.504|   21.600| 28.93|   11.550| 0.04502| 0.2923| 2.965|F   |AP        |
#> | 1587| 0.7058| 2.887|   24.450| 36.11|   12.510| 0.05262| 0.2440| 3.371|F   |AP        |
#> | 1632| 0.3002| 2.546|   11.790| 21.93|    8.615| 0.03285| 0.2924| 3.075|F   |AP        |
#> | 1734| 0.2713| 2.543|   10.670| 18.68|    7.345| 0.04250| 0.4467| 2.044|F   |AP        |
#> | 1735| 0.1837| 2.314|    7.940| 12.06|    5.212| 0.04618| 0.4795| 1.902|F   |AP        |
#> | 1850| 0.2674| 3.012|    8.876| 17.19|    5.708| 0.04585| 0.4269| 2.115|F   |AP        |
#> | 1854| 0.2651| 2.471|   10.730| 18.05|    7.305| 0.03826| 0.4156| 2.203|F   |AP        |
#> | 1970| 0.3510| 3.076|   11.410| 21.91|    7.123| 0.04478| 0.4120| 2.189|F   |AP        |
#> | 1971| 0.3113| 2.707|   11.500| 20.83|    7.696| 0.03943| 0.4182| 2.185|F   |AP        |
#> | 2056| 0.2984| 2.444|   12.210| 17.55|    7.179| 0.04606| 0.3970| 2.257|F   |AP        |
#> | 2057| 0.2544| 2.124|   11.980| 15.37|    7.238| 0.04832| 0.3400| 2.575|F   |AP        |
#> | 2059| 0.2509| 2.542|    9.871| 14.60|    5.742| 0.05000| 0.4610| 1.957|F   |AP        |
#> | 2251| 0.2874| 2.442|   11.770| 17.56|    7.191| 0.04689| 0.4206| 2.139|F   |AP        |
#> | 2252| 0.3257| 2.810|   11.590| 18.71|    6.661| 0.05470| 0.4209| 2.103|F   |AP        |
#> | 2255| 0.3054| 2.772|   11.020| 17.81|    6.426| 0.05125| 0.4363| 2.051|F   |AP        |
#> | 1633| 0.4162| 2.645|   15.740| 25.57|    9.669| 0.04375| 0.3205| 2.746|F   |Vehicle   |
#> | 1634| 0.4738| 2.629|   18.020| 24.71|    9.402| 0.05170| 0.3114| 2.754|F   |Vehicle   |
#> | 1692| 0.1933| 2.678|    7.219| 14.24|    5.315| 0.03804| 0.4628| 1.997|F   |Vehicle   |
#> | 1693| 0.1067| 2.138|    4.989|  8.30|    3.882| 0.03701| 0.4691| 1.976|F   |Vehicle   |
#> | 1694| 0.2609| 2.006|   13.010| 18.34|    9.140| 0.03780| 0.2954| 3.001|F   |Vehicle   |
#> | 1695| 0.2408| 2.512|    9.587| 15.91|    6.335| 0.04141| 0.4384| 2.084|F   |Vehicle   |
#> | 1736| 0.2073| 2.486|    8.339| 14.05|    5.652| 0.04307| 0.4740| 1.934|F   |Vehicle   |
#> | 1846| 0.2419| 2.655|    9.110| 16.45|    6.193| 0.04327| 0.4464| 2.042|F   |Vehicle   |
#> | 1851| 0.3348| 2.684|   12.470| 20.39|    7.598| 0.04540| 0.3500| 2.529|F   |Vehicle   |
#> | 1855| 0.3879| 2.889|   13.420| 28.12|    9.732| 0.03724| 0.3651| 2.485|F   |Vehicle   |
#> | 1856| 0.2304| 2.169|   10.620| 16.81|    7.748| 0.03825| 0.3843| 2.367|F   |Vehicle   |
#> | 1968| 0.4693| 3.079|   15.240| 27.54|    8.945| 0.04239| 0.3424| 2.599|F   |Vehicle   |
#> | 1972| 0.4397| 2.889|   15.220| 27.94|    9.671| 0.03953| 0.3191| 2.789|F   |Vehicle   |
#> | 1973| 0.7865| 2.919|   26.950| 37.90|   12.980| 0.05092| 0.2696| 3.120|F   |Vehicle   |
#> | 2061| 0.2922| 2.737|   10.680| 17.71|    6.470| 0.04249| 0.4665| 1.965|F   |Vehicle   |
#> | 2064| 0.2575| 2.738|    9.404| 15.39|    5.620| 0.04884| 0.4372| 2.058|F   |Vehicle   |
#> 
#> 
#> 
#> 
#> |   AS|  Tb.BV| Tb.TV| Tb.BV/TV| Tb.BS| Tb.BS/TV|   Tb.Th|  Tb.Sp|  Tb.N|Sex |Treatment |
#> |----:|------:|-----:|--------:|-----:|--------:|-------:|------:|-----:|:---|:---------|
#> | 1581| 0.2144| 2.708|    7.917| 16.36|    6.042| 0.03725| 0.3289| 2.731|M   |AP        |
#> | 1582| 0.3558| 2.786|   12.770| 24.65|    8.845| 0.03625| 0.2727| 3.236|M   |AP        |
#> | 1583| 0.4205| 2.514|   16.730| 25.93|   10.310| 0.04401| 0.2496| 3.405|M   |AP        |
#> | 1584| 0.2149| 2.677|    8.025| 16.03|    5.988| 0.03893| 0.3341| 2.681|M   |AP        |
#> | 1585| 0.4547| 3.170|   14.350| 29.58|    9.332| 0.03966| 0.2472| 3.486|M   |AP        |
#> | 1586| 0.2185| 2.396|    9.120| 16.55|    6.907| 0.03732| 0.2988| 2.975|M   |AP        |
#> | 1630| 0.4102| 3.182|   12.890| 29.03|    9.124| 0.03665| 0.2639| 3.328|M   |AP        |
#> | 1690| 0.2971| 2.429|   12.230| 21.32|    8.777| 0.03704| 0.2553| 3.421|M   |AP        |
#> | 1843| 0.3192| 3.046|   10.480| 23.73|    7.793| 0.03750| 0.2977| 2.983|M   |AP        |
#> | 1844| 0.4945| 2.609|   18.950| 27.84|   10.670| 0.04743| 0.2323| 3.575|M   |AP        |
#> | 1847| 0.3289| 2.819|   11.670| 23.62|    8.379| 0.03672| 0.2816| 3.142|M   |AP        |
#> | 1848| 0.2488| 2.452|   10.150| 21.11|    8.610| 0.03167| 0.2563| 3.472|M   |AP        |
#> | 1852| 0.2374| 2.189|   10.850| 18.12|    8.279| 0.03463| 0.2654| 3.333|M   |AP        |
#> | 2054| 0.2308| 2.518|    9.168| 16.72|    6.641| 0.04058| 0.3699| 2.436|M   |AP        |
#> | 2055| 0.3356| 2.761|   12.160| 22.32|    8.085| 0.03975| 0.2833| 3.096|M   |AP        |
#> | 2058| 0.2143| 2.874|    7.457| 15.49|    5.391| 0.04114| 0.3247| 2.733|M   |AP        |
#> | 2060| 0.2843| 2.389|   11.900| 17.84|    7.466| 0.04487| 0.3092| 2.824|M   |AP        |
#> | 2250| 0.2750| 2.585|   10.640| 20.17|    7.803| 0.03427| 0.2829| 3.153|M   |AP        |
#> | 2253| 0.2664| 2.881|    9.244| 19.72|    6.845| 0.03824| 0.2881| 3.064|M   |AP        |
#> | 2254| 0.1570| 2.499|    6.281| 12.79|    5.117| 0.03847| 0.4491| 2.051|M   |AP        |
#> | 1631| 0.4033| 3.235|   12.470| 26.14|    8.082| 0.04125| 0.2951| 2.973|M   |Vehicle   |
#> | 1691| 0.3461| 2.654|   13.040| 22.77|    8.576| 0.04145| 0.2791| 3.119|M   |Vehicle   |
#> | 1696| 0.2608| 2.491|   10.470| 19.60|    7.869| 0.03476| 0.2812| 3.165|M   |Vehicle   |
#> | 1733| 0.3892| 2.644|   14.720| 26.05|    9.853| 0.04246| 0.2493| 3.428|M   |Vehicle   |
#> | 1845| 0.4652| 2.806|   16.580| 29.11|   10.370| 0.04339| 0.2306| 3.650|M   |Vehicle   |
#> | 1849| 0.2055| 2.169|    9.475| 15.96|    7.357| 0.03491| 0.2875| 3.102|M   |Vehicle   |
#> | 1853| 0.3030| 2.702|   11.210| 23.18|    8.578| 0.03511| 0.2962| 3.018|M   |Vehicle   |
#> | 1967| 0.3149| 2.848|   11.060| 26.10|    9.165| 0.03006| 0.2664| 3.373|M   |Vehicle   |
#> | 1969| 0.3564| 2.909|   12.250| 27.06|    9.302| 0.03186| 0.2558| 3.476|M   |Vehicle   |
#> | 2062| 0.3108| 2.616|   11.880| 22.12|    8.457| 0.03930| 0.2847| 3.086|M   |Vehicle   |
#> | 2063| 0.2814| 2.436|   11.550| 18.10|    7.431| 0.04381| 0.2991| 2.916|M   |Vehicle   |
#> | 2065| 0.2989| 3.241|    9.223| 19.47|    6.009| 0.04630| 0.3546| 2.494|M   |Vehicle   |
#> | 2067| 0.3208| 3.196|   10.040| 18.80|    5.881| 0.05365| 0.4122| 2.147|M   |Vehicle   |
#> | 2247| 0.4165| 2.643|   15.760| 25.90|    9.798| 0.04265| 0.2477| 3.444|M   |Vehicle   |
#> | 2248| 0.2670| 2.691|    9.922| 20.69|    7.688| 0.03479| 0.3042| 2.950|M   |Vehicle   |
#> | 2249| 0.3040| 2.572|   11.820| 23.80|    9.255| 0.03455| 0.2449| 3.578|M   |Vehicle   |
#> 
#>